SignalFx stats exporter
This commit implements a stats exporter that sends stats as metrics to
SignalFx (https://signalfx.com). It is modeled on the Stackdriver stats
exporter in that it uses a background worker thread that periodically
inspects all the views and constructs datapoints to send to SignalFx for
their corresponding metric timeseries (identified by the view name as
metric name, and tag key/value pairs as dimensions).